rclink 0132

+++ view_rclink/not_mine/lib/features/program/data/models/program_create/submit_program_request_model.dart
@@ -27,7 +27,7 @@ abstract class RoadInputModel with _$RoadInputModel {
     required String roadUID,
     double? fromSection,
     double? toSection,
-    int? inputValue,
+    double? inputValue,
   }) = _RoadInputModel;
 
   factory RoadInputModel.fromJson(Map<String, dynamic> json) =>
